+++ src/main/java/matteroverdrive/data/inventory/DestinationFilterSlot.java
@@ -19,7 +19,6 @@
 package matteroverdrive.data.inventory;
 
 import matteroverdrive.client.render.HoloIcon;
-import matteroverdrive.items.NetworkFlashDrive;
 import matteroverdrive.proxy.ClientProxy;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fml.relauncher.Side;
@@ -32,7 +31,7 @@ public class DestinationFilterSlot extends Slot {
 
     @Override
     public boolean isValidForSlot(ItemStack item) {
-        return item.getItem() instanceof NetworkFlashDrive;
+        return true;
     }
 
     @Override