+++ src/main/java/matteroverdrive/matter_network/events/MatterNetworkEventReplicate.java
@@ -25,19 +25,17 @@ import net.minecraft.item.ItemStack;
 public class MatterNetworkEventReplicate implements IMatterNetworkEvent {
     public final ItemPattern pattern;
     public int amount;
-    public long destination;
 
-    public MatterNetworkEventReplicate(ItemPattern itemPattern, int amount, long destination) {
+    public MatterNetworkEventReplicate(ItemPattern itemPattern, int amount) {
         this.pattern = itemPattern;
         this.amount = amount;
-        this.destination = destination;
     }
 
     public static class Request extends MatterNetworkEventReplicate {
         private boolean accepted;
 
-        public Request(ItemPattern itemPattern, int amount, long destination) {
-            super(itemPattern, amount, destination);
+        public Request(ItemPattern itemPattern, int amount) {
+            super(itemPattern, amount);
         }
 
         public void markAccepted() {
@@ -52,8 +50,8 @@ public class MatterNetworkEventReplicate implements IMatterNetworkEvent {
     public static class Ccomplete extends MatterNetworkEventReplicate {
         public final ItemStack itemStack;
 
-        public Ccomplete(ItemStack itemStack, ItemPattern itemPattern, int amount, long destination) {
-            super(itemPattern, amount, destination);
+        public Ccomplete(ItemStack itemStack, ItemPattern itemPattern, int amount) {
+            super(itemPattern, amount);
             this.itemStack = itemStack;
         }
     }